refactor: nest SepaDebit under Bank
