+++ src/Struct/V2/Order/PaymentSource/Bank/SepaExperienceContext.php
@@ -5,12 +5,12 @@
  * file that was distributed with this source code.
  */
 
-namespace Shopware\PayPalSDK\Struct\V2\Order\PaymentSource\SepaDebit;
+namespace Shopware\PayPalSDK\Struct\V2\Order\PaymentSource\Bank;
 
 use OpenApi\Attributes as OA;
 use Shopware\PayPalSDK\Struct\V2\Order\PaymentSource\Common\ExperienceContext;
 
-#[OA\Schema(schema: 'paypal_v2_order_payment_source_sepa_debit_sepa_experience_context')]
+#[OA\Schema(schema: 'paypal_v2_order_payment_source_bank_sepa_experience_context')]
 class SepaExperienceContext extends ExperienceContext
 {
     public function jsonSerialize(): array
@@ -18,9 +18,10 @@ class SepaExperienceContext extends ExperienceContext
         $data = parent::jsonSerialize();
 
         return \array_filter([
-            'locale' => $data['locale'] ?? null,
-            'return_url' => $data['return_url'] ?? null,
-            'cancel_url' => $data['cancel_url'] ?? null,
+            ...$data,
+            'landing_page' => null,
+            'shipping_preference' => null,
+            'user_action' => null,
         ]);
     }
 }
